requirements.txt
Flask<2.0.0
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Bcrypt
Flask-WTF
ldap3 # TODO - Replace with Flask ldap3 Login
gunicorn
Flask-Crontab
packaging
pywebpush
markupsafe==2.0.1
msal>=1.18,<2